+++ apps/wallet/src/components/qr-code-scanner/reader/QrCodeReaderCapacitor.tsx
@@ -1,10 +1,14 @@
 import MessageContext from '@/components/app/messageContext';
 import { QrCodePropsType } from '@/types/qrcode';
-import { useContext, useEffect, useState } from 'react';
-import { BarcodeScanner, LensFacing } from '@capacitor-mlkit/barcode-scanning';
-import { Fab } from '@mui/material';
-import { CameraswitchOutlined } from '@mui/icons-material';
 import { getCameraBoxBoundary } from '@/utils/qrcode';
+import {
+  BarcodeFormat,
+  BarcodeScanner,
+  LensFacing,
+} from '@capacitor-mlkit/barcode-scanning';
+import { CameraswitchOutlined } from '@mui/icons-material';
+import { Fab } from '@mui/material';
+import { useContext, useEffect, useState } from 'react';
 
 const QrCodeReaderCapacitor = (props: QrCodePropsType) => {
   const message = useContext(MessageContext);
@@ -18,31 +22,27 @@ const QrCodeReaderCapacitor = (props: QrCodePropsType) => {
   const start = async () => {
     if (!started) {
       if (await checkPermission()) {
-        const listener = await BarcodeScanner.addListener(
-          'barcodeScanned',
-          async (result) => {
-            const boxPosition = getCameraBoxBoundary();
-            const cornerPoints = result.barcode.cornerPoints;
-            if (cornerPoints) {
-              const xs = cornerPoints.map((item) => item[0]);
-              const ys = cornerPoints.map((item) => item[1]);
-              if (
-                Math.min(...xs) < boxPosition.left ||
-                Math.max(...xs) > boxPosition.right ||
-                Math.min(...ys) < boxPosition.top ||
-                Math.max(...ys) > boxPosition.bottom
-              ) {
-                return;
-              }
-              await listener.remove();
-              await stop();
-              props.handleScan(result.barcode.rawValue);
+        await BarcodeScanner.addListener('barcodeScanned', async (result) => {
+          const boxPosition = getCameraBoxBoundary();
+          const cornerPoints = result.barcode.cornerPoints;
+          if (cornerPoints) {
+            const xs = cornerPoints.map((item) => item[0]);
+            const ys = cornerPoints.map((item) => item[1]);
+            if (
+              Math.min(...xs) < boxPosition.left ||
+              Math.max(...xs) > boxPosition.right ||
+              Math.min(...ys) < boxPosition.top ||
+              Math.max(...ys) > boxPosition.bottom
+            ) {
+              return;
             }
-          },
-        );
+            props.handleScan(result.barcode.rawValue);
+          }
+        });
         document.body.classList.add('barcode-scanner-active');
         await BarcodeScanner.startScan({
           lensFacing: usedCamera,
+          formats: [BarcodeFormat.QrCode],
         });
         setStarted(true);
       } else {
@@ -57,6 +57,7 @@ const QrCodeReaderCapacitor = (props: QrCodePropsType) => {
     BarcodeScanner.stopScan().then(() =>
       BarcodeScanner.startScan({
         lensFacing: newFace,
+        formats: [BarcodeFormat.QrCode],
       }),
     );
     setUsedCamera(newFace);
@@ -65,6 +66,7 @@ const QrCodeReaderCapacitor = (props: QrCodePropsType) => {
   const stop = async () => {
     document.body.classList.remove('barcode-scanner-active');
     await BarcodeScanner.stopScan();
+    await BarcodeScanner.removeAllListeners();
   };
 
   useEffect(() => {