+++ apps/wallet/src/pages/wallet-page/address/AddressViewCard.tsx
@@ -1,17 +1,20 @@
+import AddressCopyable from '@/components/address-copyable/AddressCopyable';
+import AssetRow from '@/components/asset-row/AssetRow';
+import CloseIcon from '@mui/icons-material/Close';
+import EditOutlinedIcon from '@mui/icons-material/EditOutlined';
 import React from 'react';
-import { Box, Card, Divider, IconButton, Typography } from '@mui/material';
-import { Close, EditOutlined } from '@mui/icons-material';
+import { Box, Button, Divider, IconButton, Typography } from '@mui/material';
 import QRCodeSVG from '@/components/display-qrcode/QrCodeSVG';
-import CopyToClipboardIcon from '@/components/copy-to-clipboard/CopyToClipboardIcon';
 import Heading from '@/components/heading/Heading';
 import ListController from '@/components/list-controller/ListController';
-import { StateAddress } from '@/store/reducer/wallet';
+import { StateAddress, TokenInfo } from '@/store/reducer/wallet';
 import { ChainTypeInterface } from '@/utils/networks/interfaces';
-import AddressToken from './AddressToken';
 
 interface AddressViewCardPropsType {
   address: StateAddress;
   chain: ChainTypeInterface;
+  isDefault: boolean;
+  name: string;
   handleEdit: () => unknown;
   handleClose: () => unknown;
 }
@@ -21,59 +24,47 @@ const AddressViewCard = (props: AddressViewCardPropsType) => {
     <React.Fragment>
       <Box display="flex" mb={2}>
         <IconButton onClick={props.handleEdit}>
-          <EditOutlined />
+          <EditOutlinedIcon />
         </IconButton>
         <Typography variant="h1" textAlign="center" sx={{ p: 1 }}>
-          {props.address.name}
+          {props.name}
         </Typography>
         <IconButton onClick={props.handleClose}>
-          <Close />
+          <CloseIcon />
         </IconButton>
       </Box>
       {props.address.tokens.length > 0 ? (
         <React.Fragment>
           <Heading title="Tokens" />
           <ListController
+            loading={false}
             error={false}
             data={props.address.tokens}
-            loading={false}
-            render={(item, index) => (
-              <AddressToken
+            divider={false}
+            emptyTitle={''}
+            render={(item: TokenInfo) => (
+              <AssetRow
                 id={item.tokenId}
-                amount={BigInt(item.balance)}
-                key={index}
-                network_type={props.chain.label}
+                networkType={props.chain.label}
+                amount={item.balance}
               />
             )}
-            emptyTitle={''}
-            emptyDescription={''}
-            emptyIcon={''}
-            errorDescription={''}
-            errorTitle={''}
           />
-          <Divider sx={{ my: 2 }} />
         </React.Fragment>
-      ) : null}
-      <Box sx={{ p: 4, fontStyle: 'italic', color: 'textSecondary' }}>
+      ) : undefined}
+      <Divider sx={{ my: 2 }} />
+      <Box sx={{ p: 8, textAlign: 'center', fontStyle: 'italic' }}>
         <QRCodeSVG value={props.address.address} />
       </Box>
-      <Box>
-        <Card sx={{ display: 'flex', bgcolor: 'info.light', p: 2 }}>
-          <Typography sx={{ overflowWrap: 'anywhere' }}>
-            {props.address.address}
-          </Typography>
-          <CopyToClipboardIcon text={props.address.address} />
-        </Card>
-      </Box>
-      {props.address.path ? (
-        <Typography
-          sx={{ my: 2, overflowWrap: 'anywhere', textAlign: 'center' }}
-          variant="body2"
-          color="textSecondary"
-        >
-          Derivation path: {props.address.path}
-        </Typography>
-      ) : null}
+      <AddressCopyable address={props.address.address} />
+      <Typography
+        sx={{ my: 2, overflowWrap: 'anywhere', textAlign: 'center' }}
+        variant="body2"
+        color="textSecondary"
+      >
+        Derivation path: {"m/44'/429'/0'/00/"}
+      </Typography>
+      {!props.isDefault && <Button variant="text">Set as default</Button>}
     </React.Fragment>
   );
 };