+++ apps/wallet/src/store/reducer/wallet.ts
@@ -108,7 +108,13 @@ const updateAddressBalance = (
     amount: '0',
   };
   address.balance = addressBalances.amount;
-  address.tokens = addressBalances.tokens;
+  address.tokens = addressBalances.tokens.filter((item) => {
+    try {
+      return BigInt(item.balance) > 0n;
+    } catch (e) {
+      return false;
+    }
+  });
 };
 
 const walletSlice = createSlice({