+++ cli/build.sh
@@ -1,0 +1,1 @@
+npm --prefix packages/icons run build