+++ packages/icons/src/types.ts
@@ -1,0 +1,12 @@
+export type TokenType = {
+  id: string;
+  boxId: string;
+  decimals: number;
+  description: string;
+  emissionAmount: string;
+  height: number;
+  name: string;
+  networkType: string;
+  txId: string;
+  fileExtension: string;
+};