+++ apps/wallet/CHANGELOG.md
@@ -1,9 +1,14 @@
 # minotaur-wallet
 
-## 2.3.0
+## 2.4.0
 
 ### Minor Changes
 
+- Display Token Detail on transaction signing page
+- Allow multiple token burning on BurnToken dApp
+- Update address asset details page to display token icon and id and name
+- Display Token Detail in transaction generate page
+- Update Transaction Box Display page to display token id and icon
 - Display Transaction Details
 - Display all values comma separated
 - Animated qrcode display and scanning
@@ -11,7 +16,11 @@
 
 ### Patch Changes
 
+- Do not display token with amount equal to zero
+- Fix Bug in Scanner
 - Fix token select order
+- Updated dependencies
+  - @minotaur-ergo/icons@1.0.0
 
 ## 2.2.0